+++ assets/icon.stories-DYq4cdeD.js
@@ -1,0 +1,46 @@
+import{j as n}from"./iframe-D99OdOFG.js";import{A as I,C as p,G as r,S as i}from"./CacheIcon-BzW4h9Nd.js";import{A as C,V as l,W as d}from"./WebsiteIcon-KzGPuQxc.js";import{A as u}from"./AddIcon-BWP1WOYU.js";import{A as g,a as f,b as h,c as y,B as S,d as x,e as b,C as T,f as v,g as L,h as k,i as P,j as D,k as E,D as F,L as A,l as U,M as B,F as w,m as M,n as R,G as j,H,o as z,p as O,I as W,q as N,r as V,s as G,t as q,u as K,P as _,v as Q,w as X,x as Y,N as Z,y as J,z as $,O as nn,E as on,J as en,K as an,Q as mn,R as tn,S as cn,T as sn,U as In,V as pn,W as rn,X as Cn,Y as ln,Z as dn,_ as un,$ as gn,a0 as fn,a1 as hn,a2 as yn,a3 as Sn,a4 as xn,a5 as bn,a6 as Tn,a7 as vn,a8 as Ln,a9 as kn,aa as Pn,ab as Dn,ac as En,ad as Fn,ae as An,af as Un,ag as Bn,ah as wn,ai as Mn,aj as Rn,ak as jn,al as Hn,am as zn,an as On,ao as Wn}from"./DragIcon-xNRKxmNI.js";import{B as Nn}from"./BackIcon-CaobY9uL.js";import{B as Vn,F as Gn,K as qn}from"./KeyIcon-CcCZBHiJ.js";import{C as Kn,L as _n}from"./LearningIcon-BOcevj9i.js";import{C as Qn}from"./ChangelogIcon-B8loo56v.js";import{C as Xn}from"./CheckmarkIcon-BZ-AnxX1.js";import{C as Yn}from"./CloseIcon-YjqU7TmS.js";import{C as Zn,E as Jn}from"./EditIcon-D_g_efV7.js";import{D as $n}from"./TrashIcon-DVerCAVK.js";import{T as no,E as oo,a as eo,I as ao,L as mo,b as to,W as co}from"./WidowFixerIcon-BGyjIPl-.js";import{D as so,H as Io}from"./HomeIcon--Qki-M9s.js";import{E as po}from"./EmailIcon-BVvkw1Vy.js";import{F as ro}from"./FaqIcon-C4CbmEvQ.js";import{I as io}from"./ImportExportIcon-CUpXJLzV.js";import{I as Co}from"./InfoIcon-DfH8TFq5.js";import{I as lo,S as uo,T as go,U as fo}from"./UnderlinedIcon-DpS4d9Ue.js";import{R as ho}from"./ResetIcon-DzKihbTm.js";import{S as yo,U as So}from"./UpdatesIcon-DOcWzKoo.js";import{S as xo}from"./SearchIcon-B10KXk0A.js";import{S as bo}from"./SettingsIcon-DfhpmE_H.js";import{S as To}from"./SupportIcon-B2ye88Le.js";import{U as vo}from"./UsersIcon-DOLEuc49.js";import{E as Lo}from"./EyeIcon-DIZdkWAN.js";import{O as ko}from"./OpenIcon-BEfQEJZn.js";import{C as Po,a as Do}from"./ChevronRightIcon-BtQMMOEE.js";import{D as Eo}from"./DropdownIcon-Dci262fs.js";import{W as Fo,S as Ao}from"./StopIcon-WzlLx1pI.js";import"./type-Bc9hys3F.js";const Uo=[{name:"AccessibilityIcon",Component:I},{name:"ActionIcon",Component:C},{name:"AddIcon",Component:u},{name:"AdminBarIcon",Component:g},{name:"AiIcon",Component:f},{name:"AltTextIcon",Component:h},{name:"AssignedIcon",Component:y},{name:"BackIcon",Component:Nn},{name:"BackToTopIcon",Component:S},{name:"BoldIcon",Component:x},{name:"BrowserIcon",Component:b},{name:"BulletListIcon",Component:Vn},{name:"CacheIcon",Component:p},{name:"CalendarIcon",Component:T},{name:"ChangeIcon",Component:Kn},{name:"ChangelogIcon",Component:Qn},{name:"CheckmarkIcon",Component:Xn},{name:"ChildThemeIcon",Component:v},{name:"CloseIcon",Component:Yn},{name:"CloudUpIcon",Component:L},{name:"ColorsIcon",Component:k},{name:"ComingSoonIcon",Component:P},{name:"CopyIcon",Component:Zn},{name:"CustomCodeIcon",Component:D},{name:"CustomFieldsIcon",Component:E},{name:"TrashIcon",Component:$n},{name:"TopBarIcon",Component:no},{name:"DesktopIcon",Component:F},{name:"LibraryIcon",Component:A},{name:"DocumentationIcon",Component:so},{name:"DownloadIcon",Component:U},{name:"MenuIcon",Component:B},{name:"EditIcon",Component:Jn},{name:"ElementSettingsIcon",Component:oo},{name:"ElementActionsIcon",Component:eo},{name:"EmailIcon",Component:po},{name:"FallbackIcon",Component:w},{name:"FaqIcon",Component:ro},{name:"FeatureRequestIcon",Component:Gn},{name:"FontIcon",Component:M},{name:"FormsIcon",Component:R},{name:"GlossaryIcon",Component:r},{name:"GroupsIcon",Component:j},{name:"HeadingIcon",Component:H},{name:"HealthIcon",Component:z},{name:"HideIcon",Component:O},{name:"HomeIcon",Component:Io},{name:"ImportExportIcon",Component:io},{name:"IncrementIcon",Component:W},{name:"InfoIcon",Component:Co},{name:"InterfaceIcon",Component:N},{name:"ItalicIcon",Component:lo},{name:"ItemsCountIcon",Component:ao},{name:"KeyIcon",Component:qn},{name:"LayoutIcon",Component:mo},{name:"LearningIcon",Component:_n},{name:"LeftSidebarIcon",Component:to},{name:"LinkIcon",Component:V},{name:"LoginIcon",Component:G},{name:"MaintenanceConeIcon",Component:q},{name:"MaintenanceIcon",Component:K},{name:"PauseIcon",Component:_},{name:"ImageIcon",Component:Q},{name:"MediaLibraryIcon",Component:X},{name:"ModulesIcon",Component:Y},{name:"NotesIcon",Component:Z},{name:"NotificationsIcon",Component:J},{name:"NumberListIcon",Component:$},{name:"OwnersManualIcon",Component:nn},{name:"ParagraphIcon",Component:on},{name:"PasswordIcon",Component:en},{name:"PhasesIcon",Component:an},{name:"PhoneIcon",Component:mn},{name:"PluginListIcon",Component:tn},{name:"PinIcon",Component:cn},{name:"PresetsIcon",Component:sn},{name:"PrintIcon",Component:In},{name:"ProductsIcon",Component:pn},{name:"ReplyIcon",Component:rn},{name:"ResetIcon",Component:ho},{name:"CheckCircleIcon",Component:Cn},{name:"RolesIcon",Component:ln},{name:"SaveIcon",Component:yo},{name:"ScrollbarIcon",Component:dn},{name:"ScrollIcon",Component:un},{name:"SearchIcon",Component:xo},{name:"SettingsIcon",Component:bo},{name:"ShareIcon",Component:gn},{name:"StartupIcon",Component:i},{name:"StatusIcon",Component:fn},{name:"StrikethroughIcon",Component:uo},{name:"FilterIcon",Component:hn},{name:"SupportIcon",Component:To},{name:"TabletIcon",Component:yn},{name:"TagIcon",Component:Sn},{name:"TemplateIcon",Component:xn},{name:"TextIcon",Component:go},{name:"ToggleIcon",Component:bn},{name:"CloseCircleIcon",Component:Tn},{name:"TroubleshootingIcon",Component:vn},{name:"UnderlinedIcon",Component:fo},{name:"UpdatesIcon",Component:So},{name:"UploadIcon",Component:Ln},{name:"DashboardCustomizeIcon",Component:kn},{name:"UsersIcon",Component:vo},{name:"UtilityIcon",Component:Pn},{name:"VideoIcon",Component:l},{name:"EyeIcon",Component:Lo},{name:"OpenIcon",Component:ko},{name:"ChevronRightIcon",Component:Po},{name:"DropdownIcon",Component:Eo},{name:"CloseDropDownIcon",Component:Do},{name:"WelcomeIcon",Component:Fo},{name:"EditorIcon",Component:Dn},{name:"EnvironmentsIcon",Component:En},{name:"QuoteIcon",Component:Fn},{name:"WebsiteIcon",Component:d},{name:"WidowFixerIcon",Component:co},{name:"TooltipIcon",Component:An},{name:"SpamIcon",Component:Un},{name:"ReplaceIcon",Component:Bn},{name:"PluginIcon",Component:wn},{name:"MeasurementIcon",Component:Mn},{name:"LocationIcon",Component:Rn},{name:"ImageFavouriteIcon",Component:jn},{name:"ImageCheckIcon",Component:Hn},{name:"HighlightIcon",Component:zn},{name:"DragIcon",Component:On},{name:"VariableIcon",Component:Wn},{name:"StopIcon",Component:Ao}],ie={title:"PAC Universal Plugin UI/Icons/Outlined",parameters:{layout:"padded"},argTypes:{size:{description:"Size of the Link",control:{type:"select"},options:["xsmall","small","medium","large","xlarge"]},color:{description:"Color of the icon",control:{type:"color"},defaultValue:"var(--pac-upui-icon)"}},args:{size:"xlarge"}},o={render:e=>n.jsx("div",{style:{display:"grid",gridTemplateColumns:"repeat(5, 1fr)",gap:20,justifyContent:"center"},children:Uo.map(({name:a,Component:s})=>n.jsxs("div",{style:{display:"flex",flexDirection:"column",alignItems:"center",justifyContent:"center",gap:"10px",border:"1px solid rgb(215, 218, 221)",padding:"6px",position:"relative"},children:[n.jsx("div",{style:{display:"flex"},children:n.jsx(s,{size:e.size,color:e.color})}),n.jsx("div",{children:a})]},a))})};var m,t,c;o.parameters={...o.parameters,docs:{...(m=o.parameters)==null?void 0:m.docs,source:{originalSource:`{
+  render: args => <div style={{
+    display: "grid",
+    gridTemplateColumns: "repeat(5, 1fr)",
+    gap: 20,
+    justifyContent: "center"
+  }}>
+      {icons.map(({
+      name,
+      Component
+    }) => <div key={name} style={{
+      display: "flex",
+      flexDirection: "column",
+      alignItems: "center",
+      justifyContent: "center",
+      gap: "10px",
+      border: "1px solid rgb(215, 218, 221)",
+      padding: "6px",
+      position: "relative"
+    }}>
+          <div style={{
+        display: "flex"
+      }}>
+            <Component size={args.size} color={args.color} />
+          </div>
+          <div>{name}</div>
+          {/* <div
+            onClick={() => {
+                          navigator.clipboard.writeText("XYUZ").then(() => {
+                toast.success("Icon name copied to clipboard!");
+              }, () => {
+                toast.error("Failed to copy icon name.");
+              });
+            }}
+            style={{
+              position: "absolute",
+              top: "5px",
+              right: "5px",
+              cursor: "pointer",
+            }}
+           >
+            <CopyIcon />
+           </div> */}
+        </div>)}
+    </div>
+}`,...(c=(t=o.parameters)==null?void 0:t.docs)==null?void 0:c.source}}};const Ce=["Outlined"];export{o as Outlined,Ce as __namedExportsOrder,ie as default};