+++ assets/index-DCCbEvUK.js
@@ -1,1 +1,0 @@
-import{g as r}from"./iframe-Bjw1izuO.js";import{r as o}from"./index-BYmlidqC.js";var t=o();const m=r(t);export{m as H,t as r};